@@ -16,7 +16,7 @@ write_to = "ape_solidity/version.py"
 
 [tool.black]
 line-length = 100
-target-version = ['py38', 'py39', 'py310', 'py311']
+target-version = ['py39', 'py310', 'py311', 'py312']
 include = '\.pyi?$'
 
 [tool.pytest.ini_options]
@@ -30,7 +30,10 @@ addopts = """
 """
 python_files = "test_*.py"
 testpaths = "tests"
-markers = "fuzzing: Run Hypothesis fuzz test suite"
+markers = """
+fuzzing: Run Hypothesis fuzz test suite
+install: Tests that will install a solc version (slow)
+"""
 
 [tool.isort]
 line_length = 100
