@@ -69,15 +69,20 @@ setup(
     include_package_data=True,
     install_requires=[
         "py-solc-x>=2.0.2,<3",
-        "eth-ape>=0.7.10,<0.8",
+        "eth-ape>=0.8.4,<0.9",
         "ethpm-types",  # Use the version ape requires
         "eth-pydantic-types",  # Use the version ape requires
         "packaging",  # Use the version ape requires
         "requests",
     ],
-    python_requires=">=3.8,<4",
+    python_requires=">=3.9,<4",
     extras_require=extras_require,
     py_modules=["ape_solidity"],
+    entry_points={
+        "ape_cli_subcommands": [
+            "ape_solidity=ape_solidity._cli:cli",
+        ],
+    },
     license="Apache-2.0",
     zip_safe=False,
     keywords="ethereum",
@@ -91,7 +96,6 @@ setup(
         "Operating System :: MacOS",
         "Operating System :: POSIX",
         "Programming Language :: Python :: 3",
-        "Programming Language :: Python :: 3.8",
         "Programming Language :: Python :: 3.9",
         "Programming Language :: Python :: 3.10",
         "Programming Language :: Python :: 3.11",
